Separate out code that writes to store from the store itself
